feat: update content in index.html
Rewritten version of myself as a person in TDLR form was written in
the short description area underneath my name in the first footer of
the page, along with removing of the previous style inside of the file.
Furthermore, the copyright label content was changed from ARR (All
Rights Reversed) to regular copyright that can be reviewed in the
GitHub repo itself.